requirements.txt
pygame>=2.0
mediapipe>=0.10
opencv-python>=4.0
numpy>=1.20
Pillow>=10.0
openai>=1.0
python-dotenv>=1.0
